client: Add a generic request filter.
This new, client level filter allows to block or unblock "all" requests
with a specific context. It is higher level than the raw
`add_request_filter`, which itself is higher level than disconnecting
random signal handlers. This is the "middle step" of adding an even
higher level API to integrate this with the `awful.rules` system.

The reason this intermediate API is exposed is because it would
otherwise leave a gap for external modules or extension. It would have
been impossible to undo that the rules would try to do even with
custom rule sources (because, here, `nil` means something).