spawn: Add a `request::autostart` signal.
A long requested feature to allow programs to be started once from
`rc.lua`. It can be useful for config that start background scripts.
Using this to execute ~/.config/autostart is a possibility, but it is
not recommanded. In any case, that's outside of this commit scope and
first requires a better `.desktop` module.