Add a new `gears.watcher` module.
It mostly replaces `awful.widgets.watch` with a more generic API.
`awful.widgets.watch` was only useful for textbox (without abusing
of its low-level callbacks). This can be used both as a generic
polling module for files and commands. It can also be attached
directly to widgets or mixed with `gears.reactive` to build an
higher level output.