core: Expose the mainloop timing information to Lua.
AwesomeWM already calls `gettimeofday()` twice per event loop iteration.
Soe might argue this is overkill for what it previously acheived
(invisible performance warnings). But now it is used for something
user facing.

By exposing them to Lua and adding a "limit" to how long we can/should
expect a polling cycle to take, we can detect suspended laptops. Can
it be done using UPower or SystemD? Yes. However this isn't portable.

Since `gears.timer` "knows" how long the shortest cycle should be.
Also since cliking or moving the mouse on a wibox always trigger
an event loop anyway, we can detect if the time between the event
loop is greater than the time before the next `gears.timer` timeout.

With this data, it is possible to detect that the machine was non-
functional for one reason or another. AwesomeWM can then take
housekeeping actions, such as refreashing the `textclock`.

This commit only contains the C part. Future commits will implement
the other side of this feature.