timer: Refactor before adding more properties.
The next few commit will significantly extend the timer
API to bring it closer to a "cron like" module. The goal
is to create a way forward for some long standing bugs
like the `textclock` not showing the correct time when
resuming from system suspend or not updating "at :00 each
minute".

This commit is the groundwork for those changes.