shims: Add support for client restacking.
New in-progress low-level feature to allow the Lua VM to control the
client z-index. This was previously mostly ignored by the shims because
it wasn't used by any test or images.

This modifies the `_stacking.lua` a little to avoid `type() == "drawin"`,
which is (willfully) not implemented by the shims.