shims: Fix `client.get(nil, true)`
By mistake, the order `client.get()` returned was never compatible
with the C backend. Since this is used ~350 time in the tests, it
would be very hard to change. However, since there is also a
`stacked` variant which was never implemented anyway, there was
still a way to make a "compatible" variant where it really counts.

It isn't very pretty code, but it's good enough.