client: Move the raise/lower code to Lua.
This commit is not really functional. The next commit will replace
`stack.c` with `tree.c` and move some of the logic _back_ to C.

It's not possible to fully move the `globalconf.clients` and
`globalconf.stack` without creating some subtle bugs. However, these
strutures are also ill-suited for both z-index and i3-like tree
layouts.

A well known bug/limitation of the current original implementation
was that any tags would modify the global stack and "break" other
tag layouts. This was unavoidable with `sticky` and multi-tagged
clients, but also wiboxes.

I could have fixed the papercut and make the `_stacking.lua` code
work properly, but it's a waste of time. This code is a dead end
and the `tree.c` one will replace it with a saner API.