Replace client.hidden with client.visible
First of all, this makes the API more consistent with the wibox.

But, also, this is a groundwork commit to replace the stacking
implementation in order to fix a dozen or so longstanding bugs
and feature requests.

The `visible` property mostly behave the opposite of `hidden`.
There is also a new property to expose why the client is
visible or hidden. This will be more relevant once the client
layouts gain the ability to hide clients (for example, using tabs)