Add a new CAPI `tree_node` class.
This is the C-side version of a new `awful.tree` class. This
`tree_node` class is not documented and intended for internal
use only.

It exposes a fine grained API to Lua to place and stack clients.
Those actions were technically previously possible to access
using APIs like `client:swap()`, `client.get(true)` and
`client:geometry()`, but were very hard to use and required
a large number of round trips between CAPI and Lua to
correctly position clients. This, in turn, caused flickers.

This class also add brings all client related APIs to wiboxes.
It is now possible to fix wiboxes and client in the same
layout. This means tabbing can now be bug-free rather than
"mostly usable" like all previous implementation used to be.